+++ lib/index.js
@@ -1,9 +1,12 @@
 'use strict';
 
-const { getCurrDirBase, dirExists } = require('./files');
+const { getCurrDirBase, dirExists, generateCommonFile, generateCommonFileWithDir, generateRouteFile } = require('./files');
 
 module.exports = {
   hbs: require('./hbs'),
   getCurrDirBase: getCurrDirBase,
-  dirExists: dirExists
+  dirExists: dirExists,
+  generateCommonFile: generateCommonFile,
+  generateCommonFileWithDir: generateCommonFileWithDir,
+  generateRouteFile: generateRouteFile
 }
