+++ templates/route.proxy.template.hbs
@@ -5,7 +5,7 @@ module.exports.setup = function(server, proxy) {
   /**
    * @swagger
    * {{ routeUrl }}:
-   *   {{ method }}:
+   *  {{ method }}:
    *   description: {{ swaggerConfig.description }}
    *   produces:
    *    - {{ swaggerConfig.produces }}
