+++ templates/route.template.hbs
@@ -7,7 +7,7 @@ module.exports.setup = function(server, proxy) {
   /**
    * @swagger
    * {{ environment.nonProxyPath.value }}{{ routeUrl }}:
-   *   {{ method }}:
+   *  {{ method }}:
    *   description: {{ swaggerConfig.description }}
    *   produces:
    *    - {{ swaggerConfig.produces }}
